# server/openapi.yaml
openapi: 3.0.3
info:
  title: Aa Predictor API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Healthcheck
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
  /api/calc/personal:
    post:
      summary: Calculer le nombre personnel
      description: |
        Calcule le profil numerologique d'un individu a partir de son nom et de sa date de naissance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonalPayload"
            examples:
              default:
                value:
                  firstName: Alice
                  lastName: Martin
                  date: "1994-07-16"
      responses:
        "200":
          description: Resultat personnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalResponse"
        "400":
          description: Entree invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erreur lors du calcul
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/calc/couple:
    post:
      summary: Calculer le nombre de couple
      description: |
        Calcule les profils personnels de deux partenaires, leur dynamique de couple et un score de compatibilite.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouplePayload"
            examples:
              default:
                value:
                  partnerA:
                    firstName: Alice
                    lastName: Martin
                    date: "1994-07-16"
                  partnerB:
                    firstName: Louis
                    lastName: Bernard
                    date: "1990-03-22"
      responses:
        "200":
          description: Resultat couple
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoupleResponse"
        "400":
          description: Entree invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erreur lors du calcul
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/ads/active:
    get:
      summary: Lister les publicites actives
      description: Renvoie les annonces actives et valides selon la fenetre temporelle et le flag d'activation.
      responses:
        "200":
          description: Liste des publicites actives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Advertisement"
  /api/ads:
    get:
      summary: Lister toutes les publicites (admin)
      parameters:
        - $ref: "#/components/parameters/AdminToken"
      responses:
        "200":
          description: Liste des publicites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Advertisement"
        "401":
          description: Token admin invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Creer une publicite
      parameters:
        - $ref: "#/components/parameters/AdminToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementCreate"
      responses:
        "201":
          description: Publicite creee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Advertisement"
        "400":
          description: Entree invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Token admin invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erreur serveur lors de la creation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/ads/{id}:
    patch:
      summary: Mettre a jour une publicite
      parameters:
        - $ref: "#/components/parameters/AdminToken"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementUpdate"
      responses:
        "200":
          description: Publicite mise a jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Advertisement"
        "400":
          description: Entree invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Token admin invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Publicite introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Supprimer une publicite
      parameters:
        - $ref: "#/components/parameters/AdminToken"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Suppression reussie
        "401":
          description: Token admin invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Publicite introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/ads/{id}/impression:
    post:
      summary: Enregistrer une impression publicitaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Impression enregistree
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Publicite introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/ads/{id}/click:
    post:
      summary: Enregistrer un clic publicitaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Clic enregistre
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Publicite introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/share:
    post:
      summary: Creer un lien de partage
      description: |
        Enregistre un resultat de lecture pour generer un permalien partageable. Le slug expire automatiquement une fois la date limite atteinte.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareCreateRequest"
      responses:
        "201":
          description: Lien cree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareCreateResponse"
        "400":
          description: Entree invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erreur lors de la creation du partage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/share/{slug}:
    get:
      summary: Recuperer un lien de partage
      description: |
        Retourne le payload associe a un slug valide et non expire. Renvoie `404` si introuvable, `410` si le lien est expire.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Details du partage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareDetailsResponse"
        "400":
          description: Parametre slug invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Lien introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Lien expire
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    AdminToken:
      name: X-Admin-Token
      in: header
      required: false
      schema:
        type: string
      description: Jeton admin (optionnel tant que la securite n'est pas configuree).
  schemas:
    PersonalPayload:
      type: object
      required:
        - firstName
        - lastName
        - date
      properties:
        firstName:
          type: string
          minLength: 2
          example: Alice
        lastName:
          type: string
          minLength: 2
          example: Martin
        date:
          type: string
          format: date
          description: Date de naissance au format ISO (YYYY-MM-DD)
          example: "1994-07-16"
    PartnerInput:
      allOf:
        - $ref: "#/components/schemas/PersonalPayload"
    CouplePayload:
      type: object
      required:
        - partnerA
        - partnerB
      properties:
        partnerA:
          $ref: "#/components/schemas/PartnerInput"
        partnerB:
          $ref: "#/components/schemas/PartnerInput"
    IndividualReport:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        locale:
          type: string
        number:
          type: integer
        version:
          type: integer
        updatedAt:
          type: string
          format: date
        title:
          type: string
        summary:
          type: string
        love:
          type: string
        work:
          type: string
        money:
          type: string
        health:
          type: string
        shadow:
          type: string
        overallAdvice:
          type: string
        luckyMonths:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
    CoupleReport:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        locale:
          type: string
        number:
          type: integer
        version:
          type: integer
        updatedAt:
          type: string
          format: date
        archetype:
          type: string
        dynamic:
          type: string
        strengths:
          type: array
          items:
            type: string
        blindspots:
          type: array
          items:
            type: string
        conflicts:
          type: array
          items:
            type: string
        rituals:
          type: array
          items:
            type: string
        advice:
          type: string
    PersonalResponse:
      type: object
      properties:
        number:
          type: integer
          example: 5
        report:
          $ref: "#/components/schemas/IndividualReport"
    CoupleResponse:
      type: object
      properties:
        partnerA:
          $ref: "#/components/schemas/PersonalResponse"
        partnerB:
          $ref: "#/components/schemas/PersonalResponse"
        couple:
          type: object
          properties:
            number:
              type: integer
              example: 7
            report:
              $ref: "#/components/schemas/CoupleReport"
        score:
          type: integer
          minimum: 40
          maximum: 100
          example: 84
    AdvertisementMetrics:
      type: object
      properties:
        impressions:
          type: integer
          example: 125
        clicks:
          type: integer
          example: 12
    AdvertisementBase:
      type: object
      properties:
        title:
          type: string
        imageUrl:
          type: string
          format: uri
        destinationUrl:
          type: string
          format: uri
        backgroundColor:
          type: string
          example: "#5b21b6"
        label:
          type: string
        alt:
          type: string
        isActive:
          type: boolean
          default: true
        weight:
          type: integer
          default: 1
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
    Advertisement:
      allOf:
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            metrics:
              $ref: "#/components/schemas/AdvertisementMetrics"
        - $ref: "#/components/schemas/AdvertisementBase"
    AdvertisementCreate:
      allOf:
        - $ref: "#/components/schemas/AdvertisementBase"
    AdvertisementUpdate:
      allOf:
        - $ref: "#/components/schemas/AdvertisementBase"
    ShareMetadata:
      type: object
      properties:
        channel:
          type: string
          description: Canal utilise pour partager le lien (email, sms, whatsapp, etc.)
          example: whatsapp
    ShareCreateRequest:
      type: object
      required:
        - results
      properties:
        results:
          $ref: "#/components/schemas/CoupleResponse"
        metadata:
          $ref: "#/components/schemas/ShareMetadata"
    ShareCreateResponse:
      type: object
      required:
        - slug
        - url
        - expiresAt
      properties:
        slug:
          type: string
          example: kp9Xf12dQwA3
        url:
          type: string
          format: uri
          example: https://aapredictor.com/share/kp9Xf12dQwA3
        expiresAt:
          type: string
          format: date-time
          example: 2025-11-16T12:00:00.000Z
    ShareDetailsResponse:
      type: object
      required:
        - slug
        - results
        - createdAt
        - expiresAt
      properties:
        slug:
          type: string
        results:
          $ref: "#/components/schemas/CoupleResponse"
        metadata:
          $ref: "#/components/schemas/ShareMetadata"
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    ErrorIssue:
      type: object
      properties:
        path:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ErrorIssue"